requirements.txt
# asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
# iniconfig           2.1.0
# packaging           25.0
# pip===25.0
# pluggy              1.6.0
psycopg2-binary==2.9.10
pytest==8.3.5
pytest-django==4.11.1
redis==6.1.0
# sqlparse==0.5.3
